LPS-75080 Moved WorkflowGroupServiceSettings to internal.settings
